refactor: rename duplicate file names for clarity
Rename files with duplicate base names across directories:
- lib/tauri/types.ts → tauriTypes.ts
- lib/tauri/schemas.ts → tauriSchemas.ts
- lib/tauri/commands.ts → tauriCommands.ts
- pages/ProtocolEditor/types.ts → protocolEditorTypes.ts

All internal imports updated accordingly.

Closes #89

via [Happy](https://happy.engineering)
